+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Unreleased changes
 
 * Fixed
     * Clarified license (it's BSD3)
+    * Bug in reference Reader monad implementation,
+      exposing a second bug in the local-local property.
 
 0.0.1
 -----